initial commit
this is a good foundation for future raylib stuff. I spent the last few
days learning from Raylib examples and mostly spent a lot of time
refamiliarizing myself with C compiling and makefiles.

now that compilation on windows and linux for desktop and web is nice
and easy and robust, I'm thinking I can work on game state stuff next.
I should probably just make an actual game, but I want a decent menu
even for testing. idk if it's procrastinating or perfectionism or
stupidity or what. either way I'm having a lot of fun and learning a lot
:)