added mouse support to the menu
this took many painful hours of work to get working exactly right.
of course a lot of that time was debugging simple things like the
MenuSelection happening after the MenuMovement, meaning that the cursor
would be drawn at the incorrect location for one frame. A lot of stuff
like that. But now it works exactly like I want from the little testing
I did.

The next big feature to add is sound. Time to do some research on some
nice beeps and boops