updated to play multiple beeps at once
this was more work than expected. I'm basically writing my own crappy
audio mixer from scratch for simple sine waves so that the audio buffer
doesn't get clipped or overflowed or some other weird issue involving
math that I'm not very good with (for now).

I will probably scrap this mess and just play sound files like a normal
person, but this was still nice to learn from and I could return to it
later maybe. There is a sort of charm and novelty to manually generated beeps.