added pause menu
now we desperately need some function(s) for state transitions between
ui and gamestate. the design / structure of the gamestate/uistate code
is probably bad, but I did come up with almost all of it on my own from
scratch. Like I'm not sure if it would be simpler/easier if the ui state
was within the game state? It would be easier to pass around and access,
but also less organizd? Idk.

I definitely need to look into how other games are structured and
organized after this project.